+++ stories/generalized-list-item.stories.css
@@ -109,9 +109,6 @@
   }
 }
 
-
-/* Ignore below */
-
 .mt-14 {
   margin-top: 3.5rem !important;
 }
@@ -121,7 +118,6 @@
 }
 
 .list-buttons li:hover {
-  /* .bg-light */
   background: #f8f9fa;
 }
 
@@ -132,9 +128,6 @@
 .text-spaced {
   letter-spacing: 0.2rem;
 }
-
-
-
 
 .btn {
   min-width: min-content !important;
